second commit: merge hr_recruitment_skills into hr_recruitment
