+++ addons/hr/static/tests/many2one_avatar_employee.test.js
@@ -2,7 +2,8 @@ import { click, contains, start, startServer } from "@mail/../tests/mail_test_he
 import { mountView, onRpc } from "@web/../tests/web_test_helpers";
 import { describe, expect, test } from "@odoo/hoot";
 import { queryAttribute } from "@odoo/hoot-dom";
-import { defineHrSkillModels } from "@hr_skills/../tests/hr_skills_test_helpers";
+import { defineHrSkillModels } from "./hr_skills_test_helpers";
+
 
 describe.current.tags("desktop");
 defineHrSkillModels();
