+++ addons/hr_recruitment/models/__init__.py
@@ -22,3 +22,4 @@ from . import ir_attachment
 from . import ir_ui_menu
 from . import mail_activity_plan
 from . import hr_job_platform
+from . import hr_application_skill
