+++ addons/hr_recruitment/models/hr_application_skill.py
