+++ addons/hr_recruitment/security/hr_recruitment_security.xml
@@ -102,4 +102,23 @@
         <field name="domain_force">[('plan_id.res_model', '=', 'hr.applicant')]</field>
         <field name="perm_read" eval="False"/>
     </record>
+
+    <!-- Interviewer Access Rules -->
+    <record id="hr_applicant_skill_interviewer_rule" model="ir.rule">
+        <field name="name">Applicant Skill: Interviewer</field>
+        <field name="model_id" ref="model_hr_applicant_skill"/>
+        <field name="domain_force">[
+            '|',
+                ('applicant_id.job_id.interviewer_ids', 'in', user.id),
+                ('applicant_id.interviewer_ids', 'in', user.id),
+        ]</field>
+        <field name="groups" eval="[(4, ref('hr_recruitment.group_hr_recruitment_interviewer'))]"/>
+    </record>
+
+    <record id="hr_applicant_skill_officer_rule" model="ir.rule">
+        <field name="name">Applicant Skill: Officer</field>
+        <field name="model_id" ref="model_hr_applicant_skill"/>
+        <field name="domain_force">[(1, '=', 1)]</field>
+        <field name="groups" eval="[(4, ref('hr_recruitment.group_hr_recruitment_user'))]"/>
+    </record>
 </odoo>
