+++ addons/hr_recruitment/security/ir.model.access.csv
@@ -30,3 +30,4 @@ access_talent_pool_add_applicants_interviewer,access.talent.pool.add.applicants.
 access_talent_pool_add_applicants,access.talent.pool.add.applicants,model_talent_pool_add_applicants,group_hr_recruitment_user,1,1,1,1
 access_job_add_applicants_interviewer,access.job.add.applicants.interviewer,model_job_add_applicants,group_hr_recruitment_interviewer,0,0,0,0
 access_job_add_applicants,access.job.add.applicants,model_job_add_applicants,group_hr_recruitment_user,1,1,1,1
+access_hr_applicant_skill_interviewer,access_hr_applicant_skill_interviewer,model_hr_applicant_skill,hr_recruitment.group_hr_recruitment_interviewer,1,1,1,1
