+++ addons/hr_recruitment/static/src/components/search_job_applicant_menu/search_job_applicant_menu.js
