+++ addons/hr_recruitment/tests/__init__.py
@@ -6,3 +6,4 @@ from . import test_recruitment_talent_pools
 from . import test_recruitment
 from . import test_utm
 from . import test_recruitment_interviewer
+from . import test_recruitment_skills
