+++ addons/hr_recruitment/tests/test_recruitment_skills.py
