+++ addons/hr_recruitment/views/hr_applicant_views.xml
@@ -25,18 +25,25 @@
                         name="archive_applicant" 
                         string="Refuse" class="btn btn-secondary" 
                     />
+                    <!-- <button type="object"
+                        name="action_add_to_job"
+                        string="Move to this Job Position"
+                        class="btn-secondary"/> -->
                 </header>
                 <field name="message_needaction" column_invisible="True"/>
                 <field name="last_stage_id" column_invisible="True"/>
                 <field name="date_last_stage_update" column_invisible="True"/>
                 <field name="partner_name" string="Name" readonly="1" optional="show"/>
-                <field name="priority" widget="priority" nolabel="1" optional="show"/>
+                <field name="priority" widget="priority" nolabel="1" optional="hide"/>
                 <field name="create_date" readonly="1" optional="show" options="{'show_time': False}"/>
                 <field name="email_from" readonly="1" optional="hide"/>
                 <field name="partner_phone" widget="phone" readonly="1" optional="hide"/>
+                <field name="matching_score"/>
+                <field name="matching_skill_ids" widget="many2many_tags"/>
+                <field name="missing_skill_ids" widget="many2many_tags"/>
                 <field name="job_id" optional="show" column_invisible="context.get('default_talent_pool_ids')"/>
                 <field name="user_id" widget="many2one_avatar_user" optional="show"/>
-                <field name="categ_ids" widget="many2many_tags" options="{'color_field': 'color'}" optional="show"/>
+                <field name="categ_ids" widget="many2many_tags" options="{'color_field': 'color'}" optional="hide"/>
                 <field name="stage_id" optional="show" column_invisible="context.get('default_talent_pool_ids')"/>
                 <field name="type_id" column_invisible="True"/>
                 <field name="application_status" optional="hide" invisible="application_status == 'ongoing'"/>
@@ -202,6 +209,22 @@
                             </group>
                         </group>
                     </page>
+                    <page string="Skills">
+                    <div class="row ms-2">
+                        <div class="o_hr_skills_editable o_hr_skills_group o_group_skills col-lg-5 d-flex flex-column">
+                            <field name="id" invisible="1"/>
+                            <field mode="list" nolabel="1" name="applicant_skill_ids" widget="skills_one2many"
+                                context="{'default_applicant_id': id, 'no_timeline': True}">
+                                <list>
+                                    <field name="skill_type_id" optional="hidden"/>
+                                    <field name="skill_id"/>
+                                    <field name="skill_level_id"/>
+                                    <field name="level_progress" widget="progressbar"/>
+                                </list>
+                            </field>
+                        </div>
+                    </div>
+                </page>
                 </notebook>
             </sheet>
             <div class="o_attachment_preview" groups="hr_recruitment.group_applicant_cv_display"/>
@@ -254,6 +277,7 @@
                 <field string="Applicant" name="partner_name"
                     filter_domain="['|', ('partner_name', 'ilike', self), ('email_from', 'ilike', self)]"/>
                 <field string="Email" name="email_from" filter_domain="[('email_from', 'ilike', self)]"/>
+                <field name="applicant_skill_ids"/>
                 <field name="job_id"/>
                 <field name="department_id" operator="child_of"/>
                 <field name="company_id" groups="base.group_multi_company"/>
@@ -604,6 +628,7 @@
                 <field name="job_id"/>
                 <field name="department_id" operator="child_of"/>
                 <field name="user_id"/>
+                <field name="applicant_skill_ids"/>
                 <filter string="Creation Date" name="year" date="create_date" default_period="year"/>
                 <separator/>
                 <filter string="Unassigned" name="unassigned" domain="[('user_id', '=', False)]"/>
@@ -629,6 +654,7 @@
                     <filter string="Department" name="department" context="{'group_by':'department_id'}"/>
                     <filter string="Tags" name="group_by_categ_ids" context="{'group_by':'categ_ids'}"/>
                     <filter string="Stage" name="stage" context="{'group_by':'stage_id'}" />
+                    <filter string="Skills" name="groupby_skills" context="{'group_by': 'skill_ids'}"/>
                     <separator/>
                     <filter string="Creation Date" name="creation_month" context="{'group_by':'create_date:month'}" help="Creation Date"/>
                 </group>
