+++ addons/hr_recruitment/views/hr_application_skill_views.xml
@@ -39,7 +39,7 @@
     <menuitem
         id="hr_recruitment_skill_type_menu"
         name="Skill Types"
-        action="hr_skills.hr_skill_type_action"
+        action="hr.hr_skill_type_action"
         parent="hr_recruitment.menu_hr_recruitment_config_employees"
         sequence="35"
     />
