--- addons/hr_recruitment_skills/i18n/ru.po
@@ -1,120 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* hr_recruitment_skills
-#
-# Translators:
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 17.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2023-10-26 21:56+0000\n"
-"PO-Revision-Date: 2024-01-30 15:14+0400\n"
-"Last-Translator: \n"
-"Language-Team: Russian (https://app.transifex.com/odoo/teams/41243/ru/)\n"
-"Language: ru\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=4; plural=(n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<12 || n%100>14) ? 1 : n%10==0 || (n%10>=5 && n%10<=9) || (n%100>=11 && n%100<=14)? 2 : 3);\n"
-
-#. module: hr_recruitment_skills
-#: model:ir.model,name:hr_recruitment_skills.model_hr_applicant
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__applicant_id
-msgid "Applicant"
-msgstr "Кандидат"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__create_uid
-msgid "Created by"
-msgstr "Создано"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__create_date
-msgid "Created on"
-msgstr "Создано"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__display_name
-msgid "Display Name"
-msgstr "Отображаемое имя"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__id
-msgid "ID"
-msgstr "ID"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant__is_interviewer
-msgid "Is Interviewer"
-msgstr "Интервьюер"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__write_uid
-msgid "Last Updated by"
-msgstr "Последнее обновление"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__write_date
-msgid "Last Updated on"
-msgstr "Последнее обновление"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__level_progress
-msgid "Progress"
-msgstr "Прогресс"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,help:hr_recruitment_skills.field_hr_applicant_skill__level_progress
-msgid "Progress from zero knowledge (0%) to fully mastered (100%)."
-msgstr "Продвигайтесь от нулевых знаний (0%) до полностью освоенных (100%)."
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant__skill_ids
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__skill_id
-msgid "Skill"
-msgstr "Навык"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__skill_level_id
-msgid "Skill Level"
-msgstr "Уровень квалификации"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__skill_type_id
-msgid "Skill Type"
-msgstr "Тип навыков"
-
-#. module: hr_recruitment_skills
-#: model:ir.ui.menu,name:hr_recruitment_skills.hr_recruitment_skill_type_menu
-msgid "Skill Types"
-msgstr "Типы навыков"
-
-#. module: hr_recruitment_skills
-#: model:ir.model,name:hr_recruitment_skills.model_hr_applicant_skill
-msgid "Skill level for an applicant"
-msgstr "Уровень квалификации соискателя"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant__applicant_skill_ids
-#: model_terms:ir.ui.view,arch_db:hr_recruitment_skills.hr_applicant_skill_view_form
-#: model_terms:ir.ui.view,arch_db:hr_recruitment_skills.hr_applicant_view_form
-#: model_terms:ir.ui.view,arch_db:hr_recruitment_skills.hr_applicant_view_search
-msgid "Skills"
-msgstr "Навыки"
-
-#. module: hr_recruitment_skills
-#. odoo-python
-#: code:addons/hr_recruitment_skills/models/hr_applicant_skill.py:0
-msgid "The skill %(name)s and skill type %(type)s doesn't match"
-msgstr "Навык %(name)s и тип навыка %(type)s не совпадают"
-
-#. module: hr_recruitment_skills
-#. odoo-python
-#: code:addons/hr_recruitment_skills/models/hr_applicant_skill.py:0
-msgid "The skill level %(level)s is not valid for skill type: %(type)s"
-msgstr "Уровень навыка %(level)s не действителен для типа навыка: %(type)s"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.constraint,message:hr_recruitment_skills.constraint_hr_applicant_skill__unique_skill
-msgid "Two levels for the same skill is not allowed"
-msgstr "Два уровня для одного навыка не допускаются"
