--- addons/hr_recruitment_skills/i18n/zh_TW.po
@@ -1,184 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* hr_recruitment_skills
-# 
-# Translators:
-# Wil Odoo, 2025
-# Tony Ng, 2025
-# 
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server saas~18.3\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-05-09 20:37+0000\n"
-"PO-Revision-Date: 2025-05-06 12:03+0000\n"
-"Last-Translator: Tony Ng, 2025\n"
-"Language-Team: Chinese (Taiwan) (https://app.transifex.com/odoo/teams/41243/zh_TW/)\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Language: zh_TW\n"
-"Plural-Forms: nplurals=1; plural=0;\n"
-
-#. module: hr_recruitment_skills
-#: model:ir.model,name:hr_recruitment_skills.model_hr_applicant
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__applicant_id
-msgid "Applicant"
-msgstr "申請人"
-
-#. module: hr_recruitment_skills
-#: model:ir.model,website_form_label:hr_recruitment_skills.model_hr_applicant
-msgid "Apply for a Job"
-msgstr "申請職位"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__create_uid
-msgid "Created by"
-msgstr "建立人員"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__create_date
-msgid "Created on"
-msgstr "建立於"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant__display_name
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__display_name
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_job__display_name
-msgid "Display Name"
-msgstr "顯示名稱"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_job__skill_ids
-msgid "Expected Skills"
-msgstr "期望技能"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant__id
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__id
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_job__id
-msgid "ID"
-msgstr "識別號"
-
-#. module: hr_recruitment_skills
-#: model:ir.model,name:hr_recruitment_skills.model_hr_job
-msgid "Job Position"
-msgstr "工作職位"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__write_uid
-msgid "Last Updated by"
-msgstr "最後更新者"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__write_date
-msgid "Last Updated on"
-msgstr "最後更新於"
-
-#. module: hr_recruitment_skills
-#. odoo-python
-#: code:addons/hr_recruitment_skills/models/hr_job.py:0
-msgid "Matching Applicants"
-msgstr "相符申請人"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant__matching_score
-msgid "Matching Score(%)"
-msgstr "匹配分數（%）"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant__matching_skill_ids
-msgid "Matching Skills"
-msgstr "相符技能"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant__missing_skill_ids
-msgid "Missing Skills"
-msgstr "缺漏技能"
-
-#. module: hr_recruitment_skills
-#: model_terms:ir.ui.view,arch_db:hr_recruitment_skills.crm_case_tree_view_inherit_hr_recruitment_skills
-msgid "Move to this Job Position"
-msgstr "移至此職位"
-
-#. module: hr_recruitment_skills
-#. odoo-python
-#: code:addons/hr_recruitment_skills/models/hr_job.py:0
-msgid "No Matching Applicants"
-msgstr "沒有相符申請人"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__level_progress
-msgid "Progress"
-msgstr "專案進度"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,help:hr_recruitment_skills.field_hr_applicant_skill__level_progress
-msgid "Progress from zero knowledge (0%) to fully mastered (100%)."
-msgstr "由尚未掌握(0%) 到達人等級 (100%)"
-
-#. module: hr_recruitment_skills
-#. odoo-javascript
-#: code:addons/hr_recruitment_skills/static/src/components/search_job_applicant_menu/search_job_applicant_menu.xml:0
-#: model:ir.actions.server,name:hr_recruitment_skills.action_applicant_search_applicant
-msgid "Search Matching Applicants"
-msgstr "搜尋相符申請人"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant__skill_ids
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__skill_id
-msgid "Skill"
-msgstr "技能"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__skill_level_id
-msgid "Skill Level"
-msgstr "技能級別"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant_skill__skill_type_id
-msgid "Skill Type"
-msgstr "技能類型"
-
-#. module: hr_recruitment_skills
-#: model:ir.ui.menu,name:hr_recruitment_skills.hr_recruitment_skill_type_menu
-msgid "Skill Types"
-msgstr "技能類型"
-
-#. module: hr_recruitment_skills
-#: model:ir.model,name:hr_recruitment_skills.model_hr_applicant_skill
-msgid "Skill level for an applicant"
-msgstr "申請人的技能水平"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.fields,field_description:hr_recruitment_skills.field_hr_applicant__applicant_skill_ids
-#: model_terms:ir.ui.view,arch_db:hr_recruitment_skills.hr_applicant_skill_view_form
-#: model_terms:ir.ui.view,arch_db:hr_recruitment_skills.hr_applicant_view_form
-#: model_terms:ir.ui.view,arch_db:hr_recruitment_skills.hr_applicant_view_search
-msgid "Skills"
-msgstr "技能"
-
-#. module: hr_recruitment_skills
-#. odoo-python
-#: code:addons/hr_recruitment_skills/models/hr_applicant_skill.py:0
-msgid "The skill %(name)s and skill type %(type)s doesn't match"
-msgstr "技能 %(name)s 和技能類型 %(type)s 不匹配"
-
-#. module: hr_recruitment_skills
-#. odoo-python
-#: code:addons/hr_recruitment_skills/models/hr_applicant_skill.py:0
-msgid "The skill level %(level)s is not valid for skill type: %(type)s"
-msgstr "技能級別 %(level)s 對技能類型無效：%(type)s"
-
-#. module: hr_recruitment_skills
-#: model:ir.model.constraint,message:hr_recruitment_skills.constraint_hr_applicant_skill_HrApplicantSkill__unique_skill
-msgid "Two levels for the same skill is not allowed"
-msgstr "不允許為同一技能提供兩個級別"
-
-#. module: hr_recruitment_skills
-#. odoo-python
-#: code:addons/hr_recruitment_skills/models/hr_job.py:0
-msgid ""
-"We do not have any applicants who meet the skill requirements for this job "
-"position in the database at the moment."
-msgstr "資料庫目前沒有符合此職位技能要求的申請人。"
