--- addons/hr_recruitment_skills/models/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from . import hr_applicant
-from . import hr_applicant_skill
-from . import hr_job
