--- addons/hr_recruitment_skills/views/hr_applicant_views.xml
@@ -1,87 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="hr_applicant_view_form" model="ir.ui.view">
-        <field name="name">hr.applicant.view.form.inherit.hr.recruitment.skills</field>
-        <field name="model">hr.applicant</field>
-        <field name="inherit_id" ref="hr_recruitment.hr_applicant_view_form"/>
-        <field name="arch" type="xml">
-            <notebook position="inside">
-                <page string="Skills">
-                    <div class="row ms-2">
-                        <div class="o_hr_skills_editable o_hr_skills_group o_group_skills col-lg-5 d-flex flex-column">
-                            <field name="id" invisible="1"/>
-                            <field mode="list" nolabel="1" name="applicant_skill_ids" widget="skills_one2many"
-                                context="{'default_applicant_id': id, 'no_timeline': True}">
-                                <list>
-                                    <field name="skill_type_id" optional="hidden"/>
-                                    <field name="skill_id"/>
-                                    <field name="skill_level_id"/>
-                                    <field name="level_progress" widget="progressbar"/>
-                                </list>
-                            </field>
-                        </div>
-                    </div>
-                </page>
-            </notebook>
-        </field>
-    </record>
-
-    <record id="hr_applicant_view_search_bis" model="ir.ui.view">
-        <field name="name">hr.applicant.view.search.inherit.skills.bis</field>
-        <field name="model">hr.applicant</field>
-        <field name="inherit_id" ref="hr_recruitment.hr_applicant_view_search_bis"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='email_from']" position="after">
-                <field name="applicant_skill_ids"/>
-            </xpath>
-        </field>
-    </record>
-
-    <record id="hr_applicant_view_search" model="ir.ui.view">
-        <field name="name">hr.applicant.view.search.inherit.skills</field>
-        <field name="model">hr.applicant</field>
-        <field name="inherit_id" ref="hr_recruitment.hr_applicant_view_search"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='user_id']" position="after">
-                <field name="applicant_skill_ids"/>
-            </xpath>
-            <filter name="stage" position="after">
-                <filter string="Skills" name="groupby_skills" context="{'group_by': 'skill_ids'}"/>
-            </filter>
-        </field>
-    </record>
-
-    <record id="crm_case_tree_view_inherit_hr_recruitment_skills" model="ir.ui.view">
-        <field name="name">hr.applicant.view.tree.inherit.skills</field>
-        <field name="model">hr.applicant</field>
-        <field name="inherit_id" ref="hr_recruitment.crm_case_tree_view_job"/>
-        <field name="mode">primary</field>
-        <field name="arch" type="xml">
-            <xpath expr="//list" position="attributes">
-                <attribute name="class" add="o_search_matching_applicant" separator=" "/>
-            </xpath>
-            <xpath expr="//list" position="inside">
-                <header>
-                    <button name="action_add_to_job" string="Move to this Job Position" type="object" class="btn-secondary"/>
-                </header>
-            </xpath>
-            <field name="application_status" position="attributes">
-                <attribute name="optional">hide</attribute>
-            </field>
-            <field name="priority" position="attributes">
-                <attribute name="optional">hide</attribute>
-            </field>
-            <field name="partner_phone" position="attributes">
-                <attribute name="optional">hide</attribute>
-            </field>
-            <field name="categ_ids" position="attributes">
-                <attribute name="optional">hide</attribute>
-            </field>
-            <field name="partner_name" position="after">
-                <field name="matching_score"/>
-                <field name="matching_skill_ids" widget="many2many_tags"/>
-                <field name="missing_skill_ids" widget="many2many_tags"/>
-            </field>
-        </field>
-    </record>
-</odoo>
