--- addons/hr_recruitment_skills/views/hr_job_views.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="hr_job_form_inherit_hr_recruitment_skills" model="ir.ui.view">
-        <field name="name">hr.job.view.form.inherit</field>
-        <field name="model">hr.job</field>
-        <field name="inherit_id" ref="hr.view_hr_job_form"/>
-        <field name="arch" type="xml">
-            <field name="contract_type_id" position="after">
-                <field 
-                    name="skill_ids"
-                    widget="many2many_tags"
-                    options="{'color_field': 'color'}"
-                    context="{'search_default_group_skill_type_id': 1}"
-                    />
-            </field>
-        </field>
-    </record>
-
-    <record id="action_applicant_search_applicant" model="ir.actions.server">
-        <field name="name">Search Matching Applicants</field>
-        <field name="model_id" ref="hr_recruitment.model_hr_job"/>
-        <field name="binding_model_id" ref="hr_recruitment.model_hr_job"/>
-        <field name="binding_view_types">form</field>
-        <field name="state">code</field>
-        <field name="code">action = records.action_search_matching_applicants()</field>
-    </record>
-</odoo>
